requirements.txt
flask
flask-cors
python-dotenv
requests
openai-whisper
piper-tts
onnxruntime